service abstraction moved to separate file, commented
